0, 1 and 2 system
If the feature in configuration.js is set to 0, it won't show the feature to anyone. If it's set to 1, it will show it for normal users if it's a normal feature, if it's a moderator feature (i.e. manage_penguins or verify_users) it will show it to moderators. If the feature is set to 2, the feature will be available for administrators only and administrators can be defined in Configuration.js by their ID's